+++ .browserslistrc
@@ -1,0 +1,1 @@
+>= 0.25%, not dead, not IE 11
