+++ src/_includes/header.njk
@@ -1,0 +1,9 @@
+<nav class="nav">
+  <div class="">
+    <div>
+      CK
+      {% include '../assets/images/arrow-right.svg' %}
+    </div>
+  </div>
+  <div class="nav__date js-nav-date">XYZ, XX XXXX</div>
+</nav>
