+++ src/assets/styles/partials/animations.pcss
@@ -1,0 +1,59 @@
+@keyframes vline {
+  0% {
+    transform: translateY(-100%);
+  }
+
+  60%,
+  100% {
+    transform: translateY(100vh);
+  }
+}
+
+@keyframes flicker {
+  0%,
+  100% {
+    opacity: 1;
+  }
+  33% {
+    opacity: 0.88;
+  }
+  66% {
+    opacity: 0.92;
+  }
+}
+
+@keyframes turn-on {
+  0% {
+    transform: scale(1, 0.6) translate3d(0, 0, 0);
+    filter: brightness(30);
+    overflow: hidden;
+  }
+
+  3.5% {
+    transform: scale(2) translate3d(0, 100%, 0);
+  }
+
+  3.6% {
+    transform: scale(2) translate3d(0, -100%, 0);
+  }
+
+  14% {
+    transform: scale(1.3, 0.6) translate3d(0, 100%, 0);
+    filter: brightness(30);
+  }
+
+  25% {
+    transform: scale(1) translate3d(0, 0, 0);
+    filter: contrast(0) brightness(0);
+  }
+
+  98% {
+    transform: scale(1) translate3d(0, 0, 0);
+    filter: contrast(2) brightness(2) saturate(2);
+  }
+
+  100% {
+    transform: none;
+    filter: none;
+  }
+}
