+++ src/assets/styles/partials/fonts.pcss
@@ -1,0 +1,8 @@
+@font-face {
+  font-family: 'VCR OSD Mono';
+  src: url('/assets/fonts/VCROSDMono.woff2') format('woff2'),
+    url('/assets/fonts/VCROSDMono.woff') format('woff');
+  font-weight: normal;
+  font-style: normal;
+  font-display: swap;
+}
