+++ src/assets/styles/partials/nav.pcss
@@ -1,0 +1,12 @@
+.nav {
+  position: sticky;
+  top: 0;
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 300px;
+  font-size: 26px;
+  line-height: 1.34;
+}
+
+.nav__date {
+}
