+++ src/assets/styles/partials/setup.pcss
@@ -1,0 +1,21 @@
+:root {
+  --text-primary: #aaa;
+  --text-secondary: #fff;
+  --font-size: 50px;
+  --line-height: 1.3619;
+}
+
+html {
+  font-family: 'VCR OSD Mono', monospace;
+  font-weight: normal;
+  font-style: normal;
+  color: var(--text-primary);
+  font-size: var(--font-size);
+  line-height: var(--line-height);
+  background-color: #000;
+}
+
+body {
+  margin: 0;
+  overflow-x: hidden;
+}
