--- src/debug.11ty.js
@@ -1,14 +1,0 @@
-module.exports = class Test {
-  data() {
-    return {
-      permalink: '/debug.json',
-    };
-  }
-
-  render(data) {
-    // eslint-disable-next-line global-require
-    const stringify = require('util').inspect(data);
-
-    return stringify;
-  }
-};
