+++ src/index.njk
@@ -3,31 +3,42 @@ title: Kaisermann | Web Development
 layout: layouts/base.njk
 ---
 
-<p>Hey 👋!</p>
-<p>
-  I'm Christian Kaisermann, a web developer from Rio with a information systems degree from PUC-Rio. I'm a very outgoing, curious and passionate person who loves
-  <a href="https://github.com/kaisermann" target="_blank" rel="noopener">coding</a>,
-  <a href="https://instagram.com/kiwistian" target="_blank" rel="noopener">photography, nature</a>
-  and
-  <a href="https://www.last.fm/user/chrispif" target="_blank" rel="noopener">music</a>.
-</p>
-<p>
-  I've been involved with programming since my childhood, when I started taking an interest in computers thanks to
-  <a href="https://en.wikipedia.org/wiki/Ragnarok_Online" target="_blank" rel="noopener">MMO games</a>. Let's say I didn't have the most
-  <a href="https://github.com/kaisermann/ragnarok-scripts" target="_blank" rel="noopener">common teenager years</a>.
-</p>
-<p>
-  I love to stare at a
-  <a href="https://justgetflux.com/" target="_blank" rel="noopener">not-so-bright screen</a>
-  while coding for hours on a constant state of
-  <a href="https://en.wikipedia.org/wiki/Flow_(psychology)" target="_blank" rel="noopener">Flow</a>
-  (here's a
-  <a href="https://open.spotify.com/user/12142121799/playlist/1oKT4BwzBziaqIs80UPbJq" target="_blank" rel="noopener">playlist</a>
-  to get you going).
-</p>
-<p>Currently, I work @<a href="https://www.vtex.com/" target="_blank" rel="noopener">VTEX</a>
-  as a Software Engineer. My main mission is to improve the quality of its digital products, front-end infrastructure and workflow.
-</p>
+<article class="bio">
+  <p>
+    I'm Christian Kaisermann,
+    <br>
+    a web developer from Rio with a information systems degree from PUC-Rio. I'm a very outgoing, curious and passionate person who loves
+    <a href="https://github.com/kaisermann" target="_blank" rel="noopener">coding</a>,
+    <a href="https://instagram.com/kiwistian" target="_blank" rel="noopener">photography, nature</a>
+    and
+    <a href="https://www.last.fm/user/chrispif" target="_blank" rel="noopener">music</a>.
+    <br>
+    <br>
+    I've been involved with programming since my childhood, when I started taking an interest in computers thanks to
+    <a href="https://en.wikipedia.org/wiki/Ragnarok_Online" target="_blank" rel="noopener">MMO games</bra>. Let's say I didn't have the most
+    <a href="https://github.com/kaisermann/ragnarok-scripts" target="_blank" rel="noopener">common teenager years</a>.
+    <br>
+    <br>
+    I love to stare at a
+    <a href="https://justgetflux.com/" target="_blank" rel="noopener">not-so-bright screen</a>
+    while coding for hours on a constant state of
+    <a href="https://en.wikipedia.org/wiki/Flow_(psychology)" target="_blank" rel="noopener">Flow</a>
+    (here's a
+    <a href="https://open.spotify.com/user/12142121799/playlist/1oKT4BwzBziaqIs80UPbJq" target="_blank" rel="noopener">playlist</a>
+    to get you going).
+    <br><br>
+    Currently, I work
+    <a href="https://www.vtex.com/" target="_blank" rel="noopener">@VTEX</a>
+    as a Software Engineer. My main mission is to improve the quality of its digital products, front-end infrastructure and workflow.
+    <br><br>
+    If you'd like to get in touch, feel free to send me an
+    <a href="mailto:christian@kaisermann.me">e-mail</a>
+    or maybe a
+    <a href="https://twitter.com/kiwistian" target="_blank" rel="noopener">tweet</a>
+    or, who knows, a message on
+    <a href="https://www.linkedin.com/in/kaisermann/" target="_blank" rel="noopener">linkedin</a>.
+  </p>
+</article>
 
 <h2>open source</h2>
 
@@ -54,14 +65,3 @@ layout: layouts/base.njk
     </li>
   {% endfor %}
 </ul>
-
-<h2>contat</h2>
-
-<p>
-  If you'd like to get in touch, feel free to send me an
-  <a href="mailto:christian@kaisermann.me">e-mail</a>
-  or maybe a
-  <a href="https://twitter.com/kiwistian" target="_blank" rel="noopener">tweet</a>
-  or, who knows, a message on
-  <a href="https://www.linkedin.com/in/kaisermann/" target="_blank" rel="noopener">linkedin</a>.
-</p>
